Consolidate plugin sources
