--- libplug/OccupancyMatrixPlugin.h
@@ -1,121 +1,0 @@
-#ifndef OCCUPANCYMATRIXPLUGIN_H
-#define OCCUPANCYMATRIXPLUGIN_H
-
-#include <stdexcept>
-#include <string>
-#include <vector>
-
-#include <nlohmann/json.hpp>
-
-#include "AnalysisLogger.h"
-#include "HistogramCut.h"
-#include "IAnalysisPlugin.h"
-#include "PlotCatalog.h"
-#include "Selection.h"
-
-namespace analysis {
-
-class OccupancyMatrixPlugin : public IAnalysisPlugin {
-  public:
-    struct PlotConfig {
-        std::string x_variable;
-        std::string y_variable;
-        std::string region;
-        Selection selection;
-        std::string output_directory = "plots";
-        std::vector<Cut> x_cuts;
-        std::vector<Cut> y_cuts;
-    };
-
-    explicit OccupancyMatrixPlugin(const nlohmann::json &cfg) {
-        if (!cfg.contains("occupancy_matrix_plots") ||
-            !cfg.at("occupancy_matrix_plots").is_array())
-            throw std::runtime_error(
-                "OccupancyMatrixPlugin missing occupancy_matrix_plots");
-        for (auto const &p : cfg.at("occupancy_matrix_plots")) {
-            PlotConfig pc;
-            pc.x_variable = p.at("x").get<std::string>();
-            pc.y_variable = p.at("y").get<std::string>();
-            pc.region = p.at("region").get<std::string>();
-            pc.output_directory =
-                p.value("output_directory", std::string("plots"));
-            if (p.contains("x_cuts")) {
-                for (auto const &c : p.at("x_cuts")) {
-                    auto dir =
-                        c.at("direction").get<std::string>() == "GreaterThan"
-                            ? CutDirection::GreaterThan
-                            : CutDirection::LessThan;
-                    pc.x_cuts.push_back({c.at("threshold").get<double>(), dir});
-                }
-            }
-            if (p.contains("y_cuts")) {
-                for (auto const &c : p.at("y_cuts")) {
-                    auto dir =
-                        c.at("direction").get<std::string>() == "GreaterThan"
-                            ? CutDirection::GreaterThan
-                            : CutDirection::LessThan;
-                    pc.y_cuts.push_back({c.at("threshold").get<double>(), dir});
-                }
-            }
-            plots_.push_back(std::move(pc));
-        }
-    }
-
-    void onInitialisation(AnalysisDefinition &def,
-                          const SelectionRegistry &) override {
-        for (auto &pc : plots_) {
-            try {
-                RegionKey rkey{pc.region};
-                pc.selection = def.region(rkey).selection();
-            } catch (const std::exception &) {
-                log::error("OccupancyMatrixPlugin::onInitialisation",
-                           "Unknown region:", pc.region);
-            }
-        }
-    }
-    void onPreSampleProcessing(const SampleKey &, const RegionKey &,
-                               const RunConfig &) override {}
-    void onPostSampleProcessing(const SampleKey &, const RegionKey &,
-                                const RegionAnalysisMap &) override {}
-
-    void onFinalisation(const AnalysisResult &result) override {
-        if (!loader_) {
-            log::error("OccupancyMatrixPlugin::onFinalisation",
-                       "No AnalysisDataLoader context provided");
-            return;
-        }
-        for (auto const &pc : plots_) {
-            RegionKey rkey{pc.region};
-            VariableKey x_key{pc.x_variable};
-            VariableKey y_key{pc.y_variable};
-            if (!result.hasResult(rkey, x_key) || !result.hasResult(rkey, y_key)) {
-                log::error("OccupancyMatrixPlugin::onFinalisation",
-                           "Missing variables for region", rkey.str());
-                continue;
-            }
-            PlotCatalog catalog(*loader_, 800, pc.output_directory);
-            catalog.generateOccupancyMatrixPlot(
-                result, pc.x_variable, pc.y_variable, pc.region,
-                pc.selection, pc.x_cuts, pc.y_cuts);
-        }
-    }
-
-    static void setLoader(AnalysisDataLoader *loader) { loader_ = loader; }
-
-  private:
-    std::vector<PlotConfig> plots_;
-    inline static AnalysisDataLoader *loader_ = nullptr;
-};
-
-}
-
-#ifdef BUILD_PLUGIN
-extern "C" analysis::IAnalysisPlugin *createPlugin(const nlohmann::json &cfg) {
-    return new analysis::OccupancyMatrixPlugin(cfg);
-}
-extern "C" void setPluginContext(analysis::AnalysisDataLoader *loader) {
-    analysis::OccupancyMatrixPlugin::setLoader(loader);
-}
-#endif
-
-#endif
